Add files via upload
this project was experimenting to understand how people had embedded droppable payloads in harmless applications.  
may contain imsorry.exe which bitblts the screen untill the keys "imsorry" are depressed in that order one after the other.
may flag as virus but there is no harmful code or code that can damage anything.  i would flag it simply on the "formless" and "dropper" files ;-)